+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows [this format](https://github.com/sensu-plugins/community/blob/master/HOW_WE_CHANGELOG.md).
 
 ## [Unreleased]
+### Changed
+- Don't escape HTML, since RocketChat doesn't support HTML formatting.
+
 
 ## 1.0.1 - 2018-11-20
 ### Added